+++ AmazingMeSample/app/src/test/java/com/amazingme/activities/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.amazingme.amazingmesample;
+package com.amazingme.activities;
 
 import org.junit.Test;
